fix: 구입 요청 검증 위치 변경
