+++ src/main/java/store/service/StoreService.java
@@ -39,8 +39,7 @@ public class StoreService {
     public void order(List<OrderRequest> orders) {
         Store store = storeRepository.get();
         Customer customer = customerRepository.get();
-        // validate
-
+        orders.forEach(request -> store.validateStocks(request.name(), request.quantity()));
         orders.forEach(request -> store.buy(customer, request.name(), request.quantity()));
         noticeRepository.saveAllByCustomer(customer);
     }
